First token metadata loading test passes
